UPD src -- add `FIND-NAME` to the kernel
Add the words `FIND-NAME-IN` and `FIND-NAME` to the kernel
(they were accepted in 2018).

All other words to search the dictionary are defined using `FIND-NAME-IN`
(directly or indirectly).
Note: changing the `SEARCH-WORDLIST` vector affects `SFIND` and `FIND`,
but does not affect `FIND-NAME`. Changing the `FIND-NAME-IN` vector
affects `SEARCH-WORDLIST` and other.

Update "lib/ext/caseins.f" accordingly.

Add the word `?FOUND ( x\0 -- x | 0 -- never )`.